+++ pallets/liquid-staking-v2/src/tests.rs
@@ -5,13 +5,13 @@ use crate::{
 };
 use frame_support::{assert_err, assert_ok, traits::Hooks};
 use orml_traits::MultiCurrency;
-use primitives::{Balance, CurrencyId, Rate};
+use primitives::{Balance, CurrencyId, Rate, TokenSymbol};
 use sp_runtime::traits::One;
 
 #[test]
 fn stake_should_work() {
     new_test_ext().execute_with(|| {
-        assert_ok!(LiquidStaking::stake(Origin::signed(Alice), 10));
+        assert_ok!(LiquidStaking::stake(Origin::signed(ALICE), 10));
         // Check storage is correct
         assert_eq!(ExchangeRate::<Test>::get(), Rate::one());
         assert_eq!(StakingPool::<Test>::get(), 10);
@@ -25,15 +25,18 @@ fn stake_should_work() {
 
         // Check balance is correct
         assert_eq!(
-            <Test as Config>::Currency::free_balance(CurrencyId::DOT, &Alice),
+            <Test as Config>::Currency::free_balance(CurrencyId::Token(TokenSymbol::DOT), &ALICE),
             90
         );
         assert_eq!(
-            <Test as Config>::Currency::free_balance(CurrencyId::xDOT, &Alice),
+            <Test as Config>::Currency::free_balance(CurrencyId::Token(TokenSymbol::xDOT), &ALICE),
             110
         );
         assert_eq!(
-            <Test as Config>::Currency::free_balance(CurrencyId::DOT, &LiquidStaking::account_id()),
+            <Test as Config>::Currency::free_balance(
+                CurrencyId::Token(TokenSymbol::DOT),
+                &LiquidStaking::account_id()
+            ),
             10
         );
     })
@@ -42,8 +45,8 @@ fn stake_should_work() {
 #[test]
 fn unstake_should_work() {
     new_test_ext().execute_with(|| {
-        assert_ok!(LiquidStaking::stake(Origin::signed(Alice), 10));
-        assert_ok!(LiquidStaking::unstake(Origin::signed(Alice), 6));
+        assert_ok!(LiquidStaking::stake(Origin::signed(ALICE), 10));
+        assert_ok!(LiquidStaking::unstake(Origin::signed(ALICE), 6));
 
         // Check storage is correct
         assert_eq!(ExchangeRate::<Test>::get(), Rate::one());
@@ -58,15 +61,18 @@ fn unstake_should_work() {
 
         // Check balance is correct
         assert_eq!(
-            <Test as Config>::Currency::free_balance(CurrencyId::DOT, &Alice),
+            <Test as Config>::Currency::free_balance(CurrencyId::Token(TokenSymbol::DOT), &ALICE),
             96
         );
         assert_eq!(
-            <Test as Config>::Currency::free_balance(CurrencyId::xDOT, &Alice),
+            <Test as Config>::Currency::free_balance(CurrencyId::Token(TokenSymbol::xDOT), &ALICE),
             104
         );
         assert_eq!(
-            <Test as Config>::Currency::free_balance(CurrencyId::DOT, &LiquidStaking::account_id()),
+            <Test as Config>::Currency::free_balance(
+                CurrencyId::Token(TokenSymbol::DOT),
+                &LiquidStaking::account_id()
+            ),
             4
         );
     })
@@ -76,7 +82,7 @@ fn unstake_should_work() {
 fn test_record_staking_settlement_ok() {
     new_test_ext().execute_with(|| {
         assert_ok!(LiquidStaking::record_staking_settlement(
-            Origin::signed(Alice),
+            Origin::signed(ALICE),
             1,
             100,
             StakingSettlementKind::Reward
@@ -90,7 +96,7 @@ fn test_record_staking_settlement_ok() {
 fn test_duplicated_record_staking_settlement() {
     new_test_ext().execute_with(|| {
         LiquidStaking::record_staking_settlement(
-            Origin::signed(Alice),
+            Origin::signed(ALICE),
             1,
             100,
             StakingSettlementKind::Reward,
@@ -99,7 +105,7 @@ fn test_duplicated_record_staking_settlement() {
 
         assert_err!(
             LiquidStaking::record_staking_settlement(
-                Origin::signed(Alice),
+                Origin::signed(ALICE),
                 1,
                 100,
                 StakingSettlementKind::Reward
@@ -117,8 +123,8 @@ enum StakeOp {
 impl StakeOp {
     fn execute(self) {
         match self {
-            Self::Stake(amount) => LiquidStaking::stake(Origin::signed(Alice), amount).unwrap(),
-            Self::Unstake(amount) => LiquidStaking::unstake(Origin::signed(Alice), amount).unwrap(),
+            Self::Stake(amount) => LiquidStaking::stake(Origin::signed(ALICE), amount).unwrap(),
+            Self::Unstake(amount) => LiquidStaking::unstake(Origin::signed(ALICE), amount).unwrap(),
         };
     }
 }
@@ -134,7 +140,7 @@ fn test_settlement_should_work() {
             (vec![], 0, (0, 0, 0), 0),
         ];
 
-        for (stake_ops, unbonding_amount, matching_result, pallet_balance) in test_case.into_iter()
+        for (stake_ops, unbonding_amount, matching_result, _pallet_balance) in test_case.into_iter()
         {
             stake_ops.into_iter().for_each(StakeOp::execute);
             assert_eq!(
@@ -142,16 +148,9 @@ fn test_settlement_should_work() {
                 matching_result
             );
             assert_ok!(LiquidStaking::settlement(
-                Origin::signed(Alice),
+                Origin::signed(ALICE),
                 unbonding_amount
             ));
-            assert_eq!(
-                <Test as Config>::Currency::free_balance(
-                    CurrencyId::DOT,
-                    &LiquidStaking::account_id()
-                ),
-                pallet_balance
-            );
             Pallet::<Test>::on_idle(0, 10000);
         }
     })